LayoutTests/platform/wk2/imported/w3c/indexeddb/abort-in-initial-upgradeneeded-expected.txt


FAIL Test that an abort in the initial upgradeneeded sets version back to 0 assert_equals: expected (undefined) undefined but got (object) object "[object IDBDatabase]"
